move skymap load functionality to dedicated class
